+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [feat]: `addSelf()` method;
 - [feat]: `scaleSelf()` method;
 - [feat]: `subSelf()` method;
+- [feat]: `rotateSelf()` method;
+- [feat]: `rotateSelf3d()` method;
 
 ## 0.10.0 (2022-11-28)